Crate portmidi

Structs

DeviceInfo
Represents what we know about a device

InputPort
Representation of an input midi port

MidiEvent
Represents a time stamped midi event. See also MidiMessage

MidiMessage
Represents a single midi message, see also MidiEvent

OutputPort
Representation of an output midi port

Enums

PortMidiError

Constants

HDRLENGTH

PM_HOST_ERROR_MSG_LEN

Functions

count_devices
Return the number of devices. This number will not change during the lifetime of the program.

get_default_input_device_id
Gets the PortMidiDeviceId for the default input, or None if there isn't one

get_default_output_device_id
Gets the PortMidiDeviceId for the default output, or None if there isn't one

get_device_info
Returns a DeviceInfo with information about a device, or None if it does not exist

get_error_text
Translate portmidi error number into human readable message. * These strings are constants (set at compile time) so client has * no need to allocate storage

get_host_error_text
Translate portmidi host error into human readable message. These strings are computed at run time, so client has to allocate storage. After this routine executes, the host error is cleared.

initialize
initialize initalizes the underlying PortMidi C library, call this before using the library.

terminate
terminate terminates the underlying PortMidi C library, call this after using the library.

Type Definitions

PortMidiDeviceId
Used by PortMidi to refer to a Midi device

PortMidiResult